requirements.txt
discord
asyncpraw
python-dotenv